requirements.txt
kivy
httpx
websockets
pillow
